+++ routes/v2/guardians/guardians.js
@@ -2,11 +2,12 @@ const express = require('express')
 const router = express.Router()
 const hash = require('../../../utils/misc/hash')
 const httpError = require('../../../utils/http-errors')
+const { httpErrorHandler } = require('../../../utils/http-error-handler.js')
 const passport = require('passport')
 const Promise = require('bluebird')
 const sequelize = require('sequelize')
 const ValidationError = require('../../../utils/converter/validation-error')
-const hasRole = require('../../../middleware/authorization/authorization').hasRole
+const { hasRole } = require('../../../middleware/authorization/authorization')
 const usersService = require('../../../services/users/users-service-legacy')
 const usersFusedService = require('../../../services/users/fused')
 const guardiansService = require('../../../services/guardians/guardians-service')
@@ -53,13 +54,10 @@ router.route('/public')
   })
 
 router.route('/:guid')
-  .get(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), hasRole(['guardianCreator', 'systemUser']), function (req, res) {
+  .get(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), function (req, res) {
     return guardiansService.getGuardianByGuid(req.params.guid)
       .then((guardian) => {
-        return guardiansService.formatGuardian(guardian)
-      })
-      .then((json) => {
-        res.status(200).send(json)
+        res.send(guardiansService.formatGuardian(guardian))
       })
       .catch(ValidationError, e => httpError(req, res, 400, null, e.message))
       .catch(sequelize.EmptyResultError, e => { httpError(req, res, 404, null, e.message) })
@@ -67,48 +65,50 @@ router.route('/:guid')
   })
 
 router.route('/:guid')
-  .patch(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), hasRole(['guardianCreator', 'systemUser']), function (req, res) {
-    const transformedParams = {}
-    const params = new Converter(req.body, transformedParams)
+  .patch(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), function (req, res) {
+    const user = req.rfcx.auth_token_info
+    const updatableBy = user.is_super || user.has_system_role ? undefined : user.id
+    const converter = new Converter(req.body, {})
 
-    params.convert('shortname').optional().toString()
-    params.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
-    params.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
-    params.convert('altitude').optional().toFloat()
-    params.convert('stream_id').optional().toString()
-    params.convert('is_visible').optional().toBoolean()
+    converter.convert('shortname').optional().toString()
+    converter.convert('latitude').optional().toFloat().minimum(-90).maximum(90)
+    converter.convert('longitude').optional().toFloat().minimum(-180).maximum(180)
+    converter.convert('altitude').optional().toFloat()
+    converter.convert('stream_id').optional().toString()
+    converter.convert('is_visible').optional().toBoolean()
 
     let mysqlTransaction, timescaleTransaction
-    return params.validate()
-      .then(async () => {
+    return converter.validate()
+      .then(async (params) => {
         mysqlTransaction = await models.sequelize.transaction()
         timescaleTransaction = await modelsTimescale.sequelize.transaction()
 
-        const guardian = await guardiansService.getGuardianByGuid(req.params.guid)
-        const updatedGuardian = await guardiansService.updateGuardian(guardian, transformedParams, { transaction: mysqlTransaction })
+        // Check the user has permission to write to stream
+        await streamsService.update(params.stream_id, {
+          latitude: params.latitude,
+          longitude: params.longitude,
+          altitude: params.altitude,
+          is_public: params.is_visible
+        }, { updatableBy, transaction: timescaleTransaction })
 
-        await streamsService.update(updatedGuardian.stream_id, {
-          latitude: transformedParams.latitude,
-          longitude: transformedParams.longitude,
-          altitude: transformedParams.altitude,
-          is_public: transformedParams.is_visible
-        }, { transaction: timescaleTransaction })
+        const guardian = await guardiansService.getGuardianByGuid(req.params.guid)
+        const updatedGuardian = await guardiansService.updateGuardian(guardian, params, { transaction: mysqlTransaction })
 
         if (arbimonService.isEnabled) {
           await arbimonService.updateSite({
-            id: updatedGuardian.stream_id,
-            name: transformedParams.shortname,
-            latitude: transformedParams.latitude,
-            longitude: transformedParams.longitude,
-            altitude: transformedParams.altitude,
-            is_private: !transformedParams.is_visible
+            id: params.stream_id,
+            name: params.shortname,
+            latitude: params.latitude,
+            longitude: params.longitude,
+            altitude: params.altitude,
+            is_private: !params.is_visible
           }, req.headers.authorization)
         }
 
         // Commit the transaction after doing update both guardian and stream
         await mysqlTransaction.commit()
         await timescaleTransaction.commit()
-        return res.status(200).send(await guardiansService.formatGuardian(updatedGuardian))
+        return res.send(guardiansService.formatGuardian(updatedGuardian))
       })
       .catch(async (e) => {
         if (mysqlTransaction) {
@@ -117,12 +117,13 @@ router.route('/:guid')
         if (timescaleTransaction) {
           await timescaleTransaction.rollback()
         }
-        httpError(req, res, 500, e, `Error while updating guardian with guid "${req.params.guid}".`); console.log(e)
+        throw e
       })
+      .catch(httpErrorHandler(req, res, 'Failed updating guardian'))
   })
 
 router.route('/register')
-  .post(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), hasRole(['rfcxUser', 'guardianCreator']), async function (req, res) {
+  .post(passport.authenticate(['token', 'jwt', 'jwt-custom'], { session: false }), async function (req, res) {
     const transformedParams = {}
     const params = new Converter(req.body, transformedParams)